Added Transformation functionality

+++ src/Engine/Clock.java
@@ -3,7 +3,7 @@ package Engine;
 public class Clock {
 
     private static long lastTime = System.nanoTime();
-    private static double deltaTime;
+    private static float deltaTime;
 
     private long startTime;
 
@@ -16,19 +16,19 @@ public class Clock {
     }
 
     // returns time since started in seconds
-    public double GetTime(){
-        return (System.nanoTime() - startTime) / 1000000000.0;
+    public float GetTime(){
+        return (System.nanoTime() - startTime) / 1000000000.0f;
     }
 
     // returns time since last frame in seconds
-    public static double DeltaTime(){
+    public static float DeltaTime(){
         return deltaTime;
     }
 
     public static void CalculateDeltaTime(){
 
         long current = System.nanoTime();
-        deltaTime = (current - lastTime) / 1000000000.0;
+        deltaTime = (current - lastTime) / 1000000000.0f;
         lastTime = current;
     }
 }
